+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 * [AMA] PDP, get count of reviews from HTML review text
 * [AMA] Search Page, remove UVP and delivery options
 * [AMA] Search Page, remove ad card on top of search results
+* [AMA] send event and item data together